Refatoração

+++ src/com/labmedicine/repository/PersonInterface.java
@@ -15,7 +15,7 @@ public interface PersonInterface<T> {
 
   List<T> getAll();
 
-  T getById(Integer indice);
+  T getById(Long indice);
 
   List<T> getByName(T object);
